bench/requirements.txt
anyio==3.6.1
fastapi==0.85.1
idna==3.4
plotly==5.10.0
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
tenacity==8.1.0
typing_extensions==4.4.0
uvicorn==0.18.3
PyYAML==6.0
